+++ busy_beaver/models/__init__.py
@@ -4,7 +4,7 @@ from .base import BaseModel  # noqa isort:skip
 # Models
 from .api_user import ApiUser  # noqa
 from .event import Event  # noqa
-from .github_summary_user import GitHubSummaryConfiguration, GitHubSummaryUser  # noqa
+from .github_summary import GitHubSummaryConfiguration, GitHubSummaryUser  # noqa
 from .key_value import KeyValueStore  # noqa
 from .slack import SlackAppHomeOpened, SlackInstallation  # noqa
 from .task import (  # noqa
