+++ busy_beaver/models/github_summary.py
